- premake working nicely. - Split MFMain() into MFMain.cpp - Added some MFString methods. - 64bit fixes.
git-svn-id: https://fuji.googlecode.com/svn/trunk@681 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Include/MFPtrList.inl
@@ -120,7 +120,7 @@ inline int MFPtrList<T>::GetLength()
 {
 	T** iterator = ppMark;
 	while(*iterator!=0) ++iterator;
-	return iterator - ppMark;
+	return (int)(iterator - ppMark);
 }
 
 template<class T>
